Use pug instead of plain html and add access workflow
Pages are saved in the 'views' directory and rendered by the server.
When a user connects, the server checks if it has a valid session token,
determines if he can access the dashboard and renders it.
If the token is not valid or doesn't exist, user is prompted to log in.