Add TTL mechanism to cache
When a user disconnects from the server, its data are not removed,
but kept in cache and its status set to offline.
In the meantime a periodic function checks the cache, decreasing
the TTL of inactive users and removing those who reached 0.
If a user connects back before expiring, its TTL is restored and its
status set to active.